+++ src/doc/html.rs
@@ -1338,6 +1338,7 @@ impl Walker {
                 alert: None,
                 range: start..start,
                 centered: false,
+                right: false,
                 details: self.details[id as usize].parent,
                 kind: BlockKind::Summary {
                     spans: vec![Span::plain("Details")],
@@ -1395,6 +1396,7 @@ impl Walker {
             alert: None,
             range,
             centered: self.center.iter().any(|&c| c),
+            right: false,
             details,
             kind,
         });